refactor to move common functions to lib
